requirements.txt
alembic==1.16.4
bcrypt==4.3.0
blinker==1.9.0
boto3==1.40.6
botocore==1.40.6
click==8.2.1
colorama==0.4.6
ffmpeg-python==0.2.0
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
future==1.0.0
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==2.3.6
WTForms==3.2.1
